auth and prisma relocation
