+++ app/(app)/admin/page.tsx
@@ -1,10 +1,10 @@
-import { auth, getUserSession } from "@/lib/auth";
+import { auth, getUserSession } from "@/lib/auth/auth";
 import { redirect } from "next/navigation";
 import { AdminStats } from "./components/admin-stats";
 import { UserList } from "./components/user-list";
 import { ShieldAlert } from "lucide-react";
 import { headers } from "next/headers";
-import { prisma } from "@/lib/prisma";
+import { prisma } from "@/lib/prisma/prisma";
 
 export default async function AdminPage() {
     const session = await getUserSession();
