+++ app/(app)/dashboard/components/cashflow-chart.tsx
@@ -1,5 +1,5 @@
-import { prisma } from "@/lib/prisma"
-import { getUserSession } from "@/lib/auth"
+import { prisma } from "@/lib/prisma/prisma"
+import { getUserSession } from "@/lib/auth/auth"
 import { getUserConfig } from "@/lib/user-config"
 import { subDays, startOfDay, format, eachDayOfInterval } from "date-fns"
 import { CashflowChartClient } from "./cashflow-chart-client"
