+++ app/(app)/dashboard/components/summary-card.tsx
@@ -1,5 +1,5 @@
-import { prisma } from "@/lib/prisma"
-import { getUserSession } from "@/lib/auth"
+import { prisma } from "@/lib/prisma/prisma"
+import { getUserSession } from "@/lib/auth/auth"
 import { MoveDownLeft, MoveUpRight, PiggyBank } from "lucide-react"
 import StatCard from "./status-card"
 import { formatAmount } from "@/utility/transaction"
