+++ app/(app)/edit-profile/page.tsx
@@ -11,7 +11,7 @@ import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar'
 import { LoadingSwap } from '@/components/ui/loading-swap'
-import { authClient, useSession } from '@/lib/auth-client'
+import { authClient, useSession } from '@/lib/auth/auth-client'
 import { BackHeader } from '@/components/back-header'
 import { getInitials } from '@/utility/party'
 import { FooterButtons } from '@/components/footer-buttons'
