+++ app/(app)/settings/components/link-account-body.tsx
@@ -7,11 +7,10 @@ import {
     SUPPORTED_OAUTH_PROVIDER_DETAILS,
     SUPPORTED_OAUTH_PROVIDERS,
     SupportedOAuthProvider,
-} from '@/lib/o-auth-providers'
+} from '@/lib/auth/o-auth-providers'
 import { useRouter } from 'next/navigation'
-import { auth } from '@/lib/auth'
-import { authClient } from '@/lib/auth-client'
-import { Button } from '@/components/ui/button'
+import { auth } from '@/lib/auth/auth'
+import { authClient } from '@/lib/auth/auth-client'
 import { cn } from '@/lib/utils'
 
 type Account = Awaited<ReturnType<typeof auth.api.listUserAccounts>>[number]
