+++ app/(app)/settings/security/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useSession } from "@/lib/auth-client";
+import { useSession } from "@/lib/auth/auth-client";
 import { SecurityModalBody } from "../components/security-body";
 import { BackHeader } from "@/components/back-header";
 import { useRouter } from "next/navigation";
