+++ app/(auth)/reset-password/page.tsx
@@ -8,7 +8,7 @@ import { motion, AnimatePresence, Variants } from "framer-motion";
 
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { authClient } from "@/lib/auth-client";
+import { authClient } from "@/lib/auth/auth-client";
 import { envClient } from "@/lib/env.client";
 
 const containerVariants: Variants = {
