+++ app/(auth)/two-factor/components/backup-code-tab.tsx
@@ -6,7 +6,7 @@ import { ShieldCheck } from "lucide-react"
 
 import { Input } from "@/components/ui/input"
 import { Button } from "@/components/ui/button"
-import { authClient } from "@/lib/auth-client"
+import { authClient } from "@/lib/auth/auth-client"
 import { toast } from "sonner"
 
 type BackupCodeFormValues = {
