+++ components/header.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar"
-import { useSession } from "@/lib/auth-client"
+import { useSession } from "@/lib/auth/auth-client"
 import { getInitials } from "@/utility/party"
 import Image from "next/image"
 import { useRouter } from "next/navigation"
