+++ prisma/seed.ts
@@ -1,4 +1,4 @@
-import { prisma } from "@/lib/prisma"
+import { prisma } from "@/lib/prisma/prisma"
 
 // Default Data
 async function main() {
